Access the "New Contact" menuitem in the contact list context menu
Contacts are managed in the contacts tab with a context menu.
Previously the menu is only shown when clicking on an existing contact,
clicking outside of the list shows nothing, and to create a new contact
one had to go in the main menu -> Wallet -> New Contact.

Change behaviour such that when in the contact menu, clicking outside
the existing contacts brings up the choice of creating a new contact.