Increase the timeout when using a single server
Only waiting 2 seconds before giving up is not a good idea when using a
single server, especially when using a slow connection. This led to
Electrum not being able to connect at all.

Waiting only 2 seconds in the case of auto-select should be fine, given
a decent network connection, as when one server doesn't reply within 2
seconds, the next will be tried at least.