Several QR code fixes
Updated qrscanner.py to keep a reference to the QR processor. This fixes
crashes after scanning on windows. The qrscanner module should init the QR
processor before scanning. This requires the camera, which is stored in
config. That is why ScanQRTextEdits require a link to a window with access
to config.

I made a split between ScanQRTextEdits and ShowQRTextEdits, because they
function very differently. The only overlap between these classes is the
fact that they have a QR icon in the edit box. No need to use
setReadOnly(True) anymore.

Update on show_master_pub_keys with a Combobox as it doesn't make sense to
show multiple QR codes in 1 screen. It would not be useable as soon as
you'd have 3+ accounts.

Fixed the copy to clipboard of show_master_pub_keys QRcode window. It has
no parent. Clipboard does not require parent() call anymore now.

Added bitcoin URI parsing on scanning a QRcode from within the send tab.

Not included yet:
- fee update on URI scan.
- ScanQRTextEdit for Add Contact dialog