Create schedule.py
Add a plugin to schedule transaction. The transaction details will first be stored in sqlite3 database, and then set a QTimer to execute the transaction at the schedule time. It will check whether there is a expired transaction when each time open the electrum. If there is a expired transaction, it will ask user whether want to do it now. If user choose 'no', the transaction will be deleted.