Common stale interface testing and switching
A new function switch_interface_if_stale() checks if the
current interface ist stale and switches if so. It optionally
takes a suggestion of the server to switch to, which is
the previous strategy of new_blockchain_height().
It checks the protocol of the suggestion before switching.
The protocol might not match if the network has
just been restarted and the prior interfaces had queued a notification.

Small changes in behaviour:

The main loop check_interfaces(): it now also tests for lag as well
as for connection status.

As check_interfaces() has a regular staleness test, such a test in
set_parameters() and on_header() is no longer necessary.

new_blockchain_height() now uses the slightly more sophisticated
switch logic of check_interfaces() that also considers disconnected servers.

Diagnotics moved to server_is_lagging().