blockchain: compute target bits for every block
When catching up, Electrum operates on chunks of size 2016 blocks.
In Bitcoin the target bits are constant for the whole chunk.
Therefore, Electrum calls the target bits adjustment function
get_target() once a chunk. But for altcoins that have a difficulty
adjustment interval lower than 2016 blocks this needs to be changed.

This change calls get_target() for every block. get_target() exits
early to make sure that the target is only adjusted once per
interval - similar as it's done in bitcoin-core.