follow-up prev: specify rlp version
python-trezor depends on rlp (an ethereum package) that we do not need..
Newer versions of rlp introduced several new dependencies, which is even more stuff we don't need.
Some of these new dependencies are even a trouble to build on Wine, as they need vc++, and they don't distribute binary builds.

related #4338