Make address related method delegate to scripthash
Initially I thought that the caller should be responsible for maintaining the
mapping between address and scripthash. That made sense b/c callers accessed
the mapping in the network module directly. And the dictionary mainting the
mapping was accessed by different callers on potentially different threads.
Turns out that maintaining that mapping on the caller site involves quite a bit
of work. Besides using address instead of scripthash makes sense from the users
perspective. The user never thinks in terms of scripthash. The concurrent
access is not a huge deal since we're only adding items and reading singular
values.