Remove mentions interface in non network modules
We're using the direct way of sending a request. This way the method call does
not wait for a response. It's a little odd that the codebase is using two
different ways of dealing with responses. One is with callbacks or blocking
and the other is a fire and forget mode in which the responses are later tide
to callbacks via message IDs.