contrib: remove git describe workaround
`git describe` fails if no tag can be found, leading
to the whole build script to fail. This is not always
desired. To prevent `git describe` from failing in
this case the `--always` flag can be passed which
causes a short commit hash to be output when a tag
is not present.