gui: use existing icons_rc.py
This cleans up all sorts of custom build steps or
manual user action and is fully compatible with
pip install git+https.

The pyrcc5 command now only needs to be used by
developers that wish to update the icons_rc.py
file.

Let's stop pretending that icons_rc.py is not
available because since the move to submodules
it is ;)