Create cashshuffle_build_shuffled_utxo_list function in the console commands.
This is a starting point to build the shuffled_utxo json object in the wallet
file.  It will loop through all utxo and build the object, or append to an existing one.  It should be improved so that duplicates don't occur.

Next steps including doing that upgrade, then working on wallet.py functions that add/remove transactions so that this list can be updated in an ongoing manner.

After this, the coinchooser logic should be built to make use of this information along with a preferences setting to enable it.