Use native libsecp256k1 if available, otherwise fall back to the old slow way. (#1025)
* Pulled in the code from electrum for ecc_fast. Now.. to work on packaging for OSX and Windows

* Added OSX packaging/building for libsecp256k1.0.dylib

* Taking a stab at adding libsecp256k1.dll for windows builds -- it works!

It builds and packages successfully for:

- Windows
- MacOS

For Linux, users need to just make sure they have libsecp256k1 installed on their system and it will find that lib wherever it is in the normal library paths. 

This should also work for Android provided libsecp256k1.so is found.

This was by and large taken from Electrum -- we're doing it their way.