Fixed a minor bug related to the Address Converter tab for new users
The bug would manifest itself as the Address Converter tab being visible
while the menu item says "Show Address Converter" (as if it were
hidden).

This bug was due to some logic/boolean errors in how the code determined
to display the text.  Now it's fixed and the correct menu text is shown
based on the actual visibility of the optional tabs, regardless of
config_key state.