Optimized the Address class. Caching to_string()
Address.to_string results are now cached in an instance variable.

This sacrifices a tiny bit of memory for a huge performance boost. I
ran the python profiler and discovered Address.to_string() is called a *LOT*
in GUI code for large wallets with many addresses (in particular
to refresh the address_list in Qt).

This is of particular concern for the upcoming CashShuffle integration
where deterministic wallets will potentially have hundreds of generated
addresses (a pair of new unique addresses is created per shuffle).

This optimization is pretty good for that.

On cache hits the performance boost is 1000x.

As it turns out, a persistent set of Address instances live in memory
because the wallet holds them -- so caching the to_string() results per
Address instance makes a whole lot of sense.

In my benchmarking the Qt address_list GUI update is now 2x faster.

The tradeoff is that some additional memory is consumed by each Address
instance to maintain the tiny cache -- each one takes up roughly 2x the
amount of memory it did previously (assuming it contains a cached string
result).

Considering that this is a tiny object to begin with, and how our
challenge right now with this codebase is really performance and not memory
footprint, I consider this an ok tradeoff.