Fixed race condition regarding cashshuffle unique addresses
Now the code is rock solid with respect to that. Added a shared set
for the wallet for "reserved addresses" that are the set of
addresses that cashshuffle is actively using to shuffle
or that it has used to recently shuffle.

This prevents the race condition where an address just sent
a tx for shuffling but it's not yet in the wallet history.

It also is faster than the previous method we used.

Required adding some locking in the wallet but it's ok
as the function we've added locks to isn't called very much.