Fix #1042: address list expanding collapsed subtrees on new tx  (#1043)
See #1042 .

The code that implemented "restoring the state" for the address list was hard to read and a bit wonky. I made it so it works more generically with any hierarchy of trees and subtrees.

Restoring the "currentItem" would violate the invariant: it would sometimes cause subtrees that we wanted to stay collapsed to expand if the item is in a collapsed subtree. This fix addresses that. I'm hoping this latter "quirk" is the cause of what you guys saw in #1042.

The previous code had out-of-spec calls to Qt in that it would sometimes call setExpanded() on items that don't live in any tree (the used_item sometimes doesn't live in any tree). Not sure if that would cause a crash or not but it's definitely not a good thing. See: http://doc.qt.io/qt-5/qtreewidgetitem.html#setExpanded

This should address #1042 !