[Qt] Updated @rate_limited decorator to be more useful
Previously it was designed to be used with objects taking only 'self' as
an arg but this is less useful than it ought to be.

Now, it saves the latest invocation's *args,**kwargs to a hidden
instance attribute and applies those latest args to the collated invocation
that eventually runs off the timer.