Increase timeout when using --oneserver and --server
Usually if the Electrum client times out then it will switch to another
server which is hopefully faster. But when --oneserver and --server are used
then the client will reconnect back to the same slow server. What's needed is
for the client to tolerate any slowness if the user specifically configured
for a single server to be used.