[Qt] Added an explicit call to the GC on ElectrumWindow close
ElectrumWindow instances take forever to make their way to being garbage
collected because they create a circular reference zoo around them. They
end up stuck in "generation 2" of the collector for ages before the
collector finally realizes they need to be cleaned up.

So, I added a mechanism to schedule a single-shot gc.collect() call
after ElectrumWindows are closed in the UI. This call takes 40-100ms to
execute (depending on hardware) and frees up megabytes of memory.

Since closing a wallet window is a relatively infrequent UI event, and
the payoff in terms of memory savings is large, this trade-off seems
worth it.

Overall this should reduce the memory footprint of the app especially if
users open and close large wallets frequently.

Also, if feels good to not keep garbage in memory when we know it's not
needed. :)