[Qt] Got rid of pathological network dialog 'updated' spam
There was massive "update" event spam going on with respect to the
network dialog. The network dialog would register itself for "updated"
events directly from the network, in addition to subscribing to the
app-global "network_updated_signal" (which itself was intended to be
used for network updated events!).

This was reduntant and wasteful.  Each wallet window would emit() that
network_updated_signal whenever THEY got an 'updated' signal from the
network thread. This was redundant -- and the network dialog would get
that signal in addition to the network layer's own 'updated' signal
directly.

So a pathological situation would ensue: What would happen is the
'NetworkDialog.on_update' calls would build up and lead to a massive
backlog.  The situation would get really pathological for multiple
wallets open all  receiving the 'updated' signal from the network.

Here is a log with print_errors added:

*** NetworkDialog.on_network
... some stuff here indicating each wallet got the 'updated signal' ...
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update
*** NetworkDialog.on_update

So basically the network dialog would end up calling the "on_update"
method on itself multiple times per network "update" signal, in addition
to the wallets doing it too. This would freeze the GUI for small
stretches of time.

I got rid of this mechanism entirely.  The app-global
"network_updated_signal" was redundant and not used anyway and just took
up space and mental energy to grok when reading the code. Instead, the
network dialog relies on the network layer directly to inform it of
update events.

And to make sure this can never happen again, I also decorated the
NetworkDialog.on_update method with a @rate_limited decorator to ensure
it can never get spammed.