[Qt] Fix to exception raised if attempting to broadcast in offline mode
In offline mode the broadcast_transaction method on the main_window
(called by lots of other subsystems) was not checking if we had a
network.  It assumed we were online.  This would lead to an exception
being raised about NoneType not having a broadcast_transaction method
(obviously!)

Fixed this to tell the user they are in offline mode and they must
restart to go online.

Also, if the network is not connected we're now catching this situation
and presenting an error window telling them they are offline. This is so
that the user doesn't end up with a 30 second delay before getting an error
anyway.