[Qt] Performance fixups. Force GUI update_tabs to run from timer_actions
This reduces the number of redundant calls to update_wallet and
update_tabs and friends.

The strategy employed is to have update_wallet and update_labels set a
flag, which is checked by the periodic 0.5 second timer that runs
anyway. The real work is done by _update_labels() and _update_wallet()
which are called from timer_actions if the flag is set.

This reduces the number of redundant calls to update_tabs() and
update_labels() (on top of the rate_limited scheme we devised) and makes
the app feel more responsive when there are big synchs going on with
lots of tx's coming in.

Previous to this each updated tx would trigger a wallet update. There's
no need to do the update right away necessarily -- it's better to
enqueue it to happen when the timer_actions callback fires.

The @rate_limited scheme already handled most of this -- this is just
icing on top of that.

These changes are necessarily for the upcoming CashShuffle updates which
will have tx's being sent in the background shuffling thread and thus we
want to minimize the user's perception of "lag" when using the GUI. Thus
we defer GUI redraws/updates for later and collate them so that
redundant updates don't happen too often.