[Qt] Make tor proxy detector run continuously while Proxy tab is visible
Inspired by a recent Electrum commit. Whenver the proxy tab is visible
in the Network Dialog, we start the TorDetector thread which polls
localhost 9050 and 9150 for Tor proxies, and displays the tor proxy
checkbox. The polling happens from a separate thread every 10 seconds.

When the proxy tab is no longer visible, the thread is stopped. Polling
only happens when the proxy tab is visible and selected.