[Qt] Follow-up: Make sure tor checkbox in unchecked if no Tor
It's not clear to me that the tor checkbox should remain checked if the
Tor proxy goes away.  It's better to uncheck it explicitly if that
happens.  It gets auto-checked again if it comes back and is the same as
the current proxy. If it's not, then better for it to be unchecked.