[Qt] Rename icons_rc.py -> icons.py, add icons.py to repository (#1097)
This is in line with addressing issue #1088

- I removed icons_rc.py and all references to it from the code and other
meta files, including readme's and build scripts

- I rebuilt icons_rc.py as icons.py using pyrcc5-3.5 for PyQt 5.11.3
(current version we use for deterministic builds). This is forwards and
backwards compatible with all extant versions of PyQt5 on python3,
regardless of exact PyQt5 version or Python3 version the user uses (I
verified this is true by trying various combinations of the two).

- The new icons file, icons.py, is now added to the repository. This
removes the burden on users to constantly re-update this file manually
when running from source (particularly affects Linux users).

- We are doing away with the electrum-icons repository as a result. Wine
and MacOSX builds no longer depend on this repository. It's perfectly
deterministic to simply checkout a specific version of the main
Electron-Cash repository to build.

I hope I got everything.