[OSX] Fixed bug where system tray notifs were silent for built app.
For some reason, codesigning the .dylibs (libusb and libsecp256k1) in
the built app would cause system tray notification messages to never
appear in the built app.  The mysteries of Apple's code signing eludes
all of us.  Not signing the dylibs is ok, and the app still runs and
passes gatekeeper. However signing them caused the notifications to
fail.

So we got rid of the line in make_osx that code signs the dylibs and
everything is peachy now.  This issue likely affects Electrum too.