Wallet performance tweak for address synchronizer
The wallet.synchronize() code is called once every 100ms in the network
loop (this needs to be changed).

This code is a bit inefficient for wallets with long histories because
it would loop through the entire history of the latest addresses.

In practice this wouldn't really matter except for in pathological
cases.

I fixed it to abort the loop early and also use slightly less slow
python list ops.