[OSX] Switched PyInstaller to our patched AppleEvents+CodeSign version
Patched PyInstaller is identical to 3.4 release except it has the
following additional features:

- Allows for AppleEvents to get forwarded to Electrum slave process
(python interpreter).  This means bitcoin: URLs work even AFTER the app
is already running (at present they do not).

- Has the addition of the command-line option --osx-codesign-identity
with allows for code signing of *all* embedded binaries and libs as the
.app is built. The PR to PyInstaller has been submitted here:
https://github.com/pyinstaller/pyinstaller/pull/3991 but has not yet
beed accepted

This commit also undoes the monkey patching of pyinstaller to achieve #2
above and is much cleaner for maintainers.

Note: I am currently hosting the custom PyIntstaller.zip on my repo
here:
https://github.com/cculianu/pyinstaller/releases/tag/3.4AppleEventsCodeSign
but feel free to fork/copy this repo under the ecdsa or spesmilo
account.