[Lib] Fixes #1130 -- Request fee estimate update every 20 mins
Previously fee updates was set to every 120 minutes. This code was
merged in verbatim from Electrum in Sept. of 2017. For Bitcoin Core it
makes sense to not update the fee estimates in the wallet too frequently
as the fee situation on core tends to have very slow backlog cycles.

For Bitcoin Cash this isn't the case. Backlogs are temporary affairs in
general.  So I set the fee estimate retrieval time from the server
to once ever 20 minutes. The call to the server is cheap anyway so it
does no harm to do it more frequently.

This addresses "unexpected" behavior seen by @im_uname on testnet, of
all places, where "low fee" warnings seemed to pop up for seemingly ok
tx's.