[iOS][Qt] Fixed esoteric QR image bug. Addresses #1144
If the generated QR image data was too long (as can happen in rare
cases if user inputs a ridiculously long description in the Receive
tab), the internal qrcode library would raise DataOverflowError.  This
condition has now been handled on iOS and Qt where the widget will
display some dummy default data.

While an error window would have been better -- the way the QRWidget is
used in Qt (and similarly the qr UIImage in iOS) -- it wasn't easy to
propagate the error outward without lots of new code.

So.. we just put dummy graphics on-screen.  It's good enough for this
esoteric error.

Hopefully the user will realize they entered too much data (it's on the
order of hundreds of characters) and fix the situation.

It's better than before where internally the data would get so
inconsistent the user would be spammed with error windows...