[Lib] Fix long-standing bug where Network class modified DEFAULT_SERVERS
The server list was in NetworkConstants (now called networks.net), yet
it was being written-to by the Network class in networks.py. There was
no good reason for this and I assumed it was somehow intentional.

I noticed a recent Electrum commit finally addresses this, so I decided
to also fix this weirndess.

Network constants... should remain constant!

[Qt] I also fixed a UI quirk of the network dialog whereby the server/host
text edit was editable if "preferred only" was checked, but manually
entering a server there would not work in preferred only mode because
manually inputted servers aren't in the preferred list!

So the text edits are editable ONLY if BOTH auto-connect and preferred-only
are NOT checked.