[Lib][Qt] Fix for electrum issue we also had -- normalize symlinked wallets
Same wallet can be opened (by the daemon especially) if wallet file was
symlinked.  We should take steps to avoid this as it can cause
corruption.

Now, wallets are always normalized to a standardized path that
resolves symlinks.

Also refactored the code for generating a new wallet name into
electroncash.util as get_new_wallet_name