[Lib] Fix #1079 -- interface.blockchain is sometimes None
There are still many extant race condition related bugs in the network
code. This is a workaround to close issue #1079 which happens a lot in
the wild and has been happening for some time.

Rather than the user getting a crash report, we gracefully just
terminate the connection to the server when the code gets in this
inconsistent state, which will blow away the inconsistent state for that
interface.

It is what it is. :/