[Lib] Make network SPV use asynchronous requests
SPV class was using synchronous network calls which were spammy and
impacted GUI responsiveness. Instead, we switched over to a request
queue mechanism for the SPV verify tx. This should improve GUI
responsiveness when synching large wallets.

More work still needs to be done, though, to that end as the Verifier
eats tons of CPU and the GUI is still very sluggish whilst verifying.