[Qt] SPV performance improved drastically
The GUI would sieze up and lock while verifying wallets with huge
numbers of tx's (such a wallet is this one:
bitcoincash:pp8skudq3x5hzw8ew7vzsw8tn4k8wxsqsv0lt0mf3g)

This was due to the SPV verifier spamming the GUI thread with 'verified'
callbacks which the GUI would then proceed to use to update the
history_list as tx's became verified.

This was very slow and it would lock up the GUI due to the sheer number
of verifications coming in per second off the network.

In order to mitigate this and improve app responsiveness, we instead
batch these 'verified' requests to not occur more than once per second
in large batches.  This way, there is time every second for the GUI to
respond to user input.

On top of that, batching the updates reduces redundant repaint calls in
Qt as they are done all at once, drastically improving overall app
performance when verifying tx's.

This is only really noticeable on large wallets with thousands of tx's
(like the one mentioned above).

Also added a small performance tweak to blockchain.py, as well as a
smallish performance tweak to history_list to cache its items better in
a WeakValueDictionary rather than relying on Qt's O(n) lookup of tree
list items.