[Qt][Lib] Various fixups and refactorings
- More cleanup of the recent 'verifier' GUI speedups. Got rid of the
signal/slot mechanism for propagating the history-list-got-a-verified-tx
condition and instead we do it using an approach using less overhead

- Got rid of the notify_transactions_signal in main_window.py -- this
signal was a bit superfluous and could just lead to Qt signal/slot spam,
taking precious cycles away from the GUI event loop. Instead, we set a
flag which gets checked in the timer_actions slot.

- Updated the @rate_limited decorator to accept a new parameter:
ts_after=True, which affects how the rate is applied and helps to
guarantee minimum dead time between collated calls.

- We now cache all QIcon instances related to the status bar (CPU cycles
were being wasted each call to update_status to re-read the PNG data
and render it to QIcon instances).

- Added timestamps to the debug output (monotonically increasing seconds
from app start)

- Added some locks around the tx_notifications and tx_verifications
queues in the GUI since they are shared between the network and GUI threads.

- Fixed a GUI bug where the Seed Dialog allowed users to edit the seed text
(this makes no sense since the seed is immutable) -- made the text edit
read-only for the seed dialog.

- More speedups and fixups to the verifier.

- Added the text '[verifying %d TXs...]' to the status bar when the
verifier has more than 10 tx's it is working on verifying (helps
indicate to users that progress is being made on verify for huge
wallets).