[Qt][Lib] synchronizer 'updated' events now filtered by wallet
GUI Speedup Level Up +1!

Previous to this commit, if the synchronizer (a wallet-specific object)
detected a change to the address history of a wallet, it would send out
a network 'updated' event.  This event would be caught by *ALL* open
wallet instanes.  In cases where users have several wallets open this
meant that ALL the opened ElectrumWindows would refresh their histories
needlessly, even if just 1 wallet's GUI was the one that needed
updating.

Needless to say -- this caused stalls in the UI.  This has now been
addressed.

1. The synchronizer sends 'self.wallet' along with the 'updated' event,
so wallet windows can check for the presence of this argument and ignore
events not intended for them.

2. The network was the other class that would send 'updated' events to
indicate network status change.  These events are not wallet-specific
and do require action by all extant windows -- so if the client code
sees no arguments to the 'updated' callback, it knows to always accept
the event.