[Lib] Fix #1160 -- Verifier messing up on testnet
This was a bug I recently added due to my transtion over from
NetworkConstants to networks.net.  I incorrectly imported:

from .networks import net

Which means that net gets its own reference and never gets updated on
'set_testnet()'.  The correct way to do it is 'from . import networks'
and refer to the symbol as 'networks.net'.

Luckily this bug was not in the last release, only in git.

I wasn't very aware of the way Python does imports, apparently. Live and
learn. :)