[Lib] Nit: Move ScriptOutput 'lookup' inner function out of loop
This is a slight performance fixup (I think) because otherwise keeping
this inner function defined inside a loop (when really it has no
external dependencies to the loop) meant it would be evaluated for each
iteration.  Maybe it shaves microseconds per iteration. :P  Not a huge
deal. More of a nit.