Added better support for Electron Cash 'offline mode'
When Electron Cash is started with the --offline command-line option, no
network traffic should occur.

CashShuffle was not really following thus rule fully as it should -- it
did check the cashshuffle server and did other small nonsense that
doesn't belong in offline mode.

This has now been fixed.  CashShuffle can be enabled for Offline-Mode
wallets, which will just mean it will segregate coins in the Send tab --
but it will NOT go out to the network.

The coins in the coins tab that would have been scheduled for shuffling
show 'Offline' as their status.