[Qt][Lib] Fix #1164 - Warn if old testnet wallet; needs regen
Warn users if opening older testnet wallets that had the invalid
xpub/xprv key format.

If we encounter the old testnet wallet xkeys which are invalid, we:

- Stop the syncrhonizer (to prevent the network thread from blowing up)
- We tell the user in the Qt GUI to upgrade their testnet wallet to the
new format by regenerating from seed.

It's minimal -- but it's better than nothing.

The reason for not doing more for them (such as auto-upgrading the
wallet) is that it would be some coding work to do so and anyway this is
testnet: the hope is testnet users are power users anyway so more isn't
really needed.