[Qt] Fix #1168: Added Wallet -> Rebuild history
This is a quick and dirty but effective way to undo wallet file
corruption. See #1168.

The user can elect to rebuild the entire wallet history. The history
will be wiped from memory and from the wallet file, and the synchronizer
and verifier will kick in and rebuild the whole history.

It has been found that users that were around for the BSV/BCH fork and
that switched between chains may have seen their wallets contain invalid
coins (Amaury Sechet reported this, see #1168).

Rebuilding the whole history using this menu option is a power user
feature that perhaps some might find useful.

Tested and works with all wallet types.