CashShuffle: Fix #67 -- Add dust change to miner fee
Previous to this commit, shuffles that involved dust change outputs
(outputs below 546 sats) would fail because they would produce a tx that
was invalid for broadcasting (dust outputs).

Now, we add dust (<546 sats) for change outputs to the miner fee and
don't produce that change output.

This should reduce the number of failed shuffles due to dust change
outputs that we were seeing (some coins indeed were unshufflable as a result!)